+++ test/MintingManager.metatx.test.js
@@ -1,7 +1,7 @@
 const { ethers, upgrades } = require('hardhat');
 const { expect } = require('chai');
 
-const { ZERO_ADDRESS, TLD } = require('./helpers/constants');
+const { ZERO_ADDRESS } = require('./helpers/constants');
 const { buildExecuteFunc } = require('./helpers/metatx');
 
 describe('MintingManager (metatx)', () => {
@@ -34,10 +34,10 @@ describe('MintingManager (metatx)', () => {
   });
 
   it('should mint through forwarder', async () => {
-    const tokenId = await unsRegistry.childIdOf(TLD.WALLET, 'test-qw11');
+    const tokenId = await unsRegistry.namehash(['test-qw11', 'wallet']);
     const { req, signature } = await buildExecuteParams(
-      'mintSLD(address,uint256,string)',
-      [receiver.address, TLD.WALLET, 'test-qw11'],
+      'issueWithRecords(address,string[],string[],string[])',
+      [receiver.address, ['test-qw11', 'wallet'], [], []],
       coinbase, tokenId);
 
     await forwarder.execute(req, signature);
@@ -46,10 +46,10 @@ describe('MintingManager (metatx)', () => {
   });
 
   it('should revert forwarding when forwarder not trusted', async () => {
-    const tokenId = await unsRegistry.childIdOf(TLD.WALLET, 'test-qw11');
+    const tokenId = await unsRegistry.namehash(['test-qw11', 'wallet']);
     const { req, signature } = await buildExecuteParams(
-      'mintSLD(address,uint256,string)',
-      [receiver.address, TLD.WALLET, 'test-qw11'],
+      'issueWithRecords(address,string[],string[],string[])',
+      [receiver.address, ['test-qw11', 'wallet'], [], []],
       coinbase, tokenId);
 
     await mintingManager.setForwarder(ZERO_ADDRESS);
@@ -60,10 +60,10 @@ describe('MintingManager (metatx)', () => {
   });
 
   it('should revert execution when signature is not valid', async () => {
-    const tokenId = await unsRegistry.childIdOf(TLD.WALLET, 'test-qw1341');
+    const tokenId = await unsRegistry.namehash(['test-qw1341', 'wallet']);
     const { req, signature } = await buildExecuteParams(
-      'mintSLD(address,uint256,string)',
-      [receiver.address, TLD.WALLET, 'test-qw1341'],
+      'issueWithRecords(address,string[],string[],string[])',
+      [receiver.address, ['test-qw1341', 'wallet'], [], []],
       coinbase, tokenId);
 
     await expect(forwarder.execute({ ...req, from: receiver.address }, signature)).to.be
@@ -71,10 +71,10 @@ describe('MintingManager (metatx)', () => {
   });
 
   it('should revert execution when used signature', async () => {
-    const tokenId = await unsRegistry.childIdOf(TLD.WALLET, 'test-qw1341');
+    const tokenId = await unsRegistry.namehash(['test-qw1341', 'wallet']);
     const { req, signature } = await buildExecuteParams(
-      'mintSLD(address,uint256,string)',
-      [receiver.address, TLD.WALLET, 'test-qw1341'],
+      'issueWithRecords(address,string[],string[],string[])',
+      [receiver.address, ['test-qw1341', 'wallet'], [], []],
       coinbase, tokenId);
 
     await forwarder.execute(req, signature);