+++ test/UNSRegistry.metatx.test.js
@@ -2,7 +2,7 @@ const { ethers } = require('hardhat');
 const { expect } = require('chai');
 
 const { sign, buildExecuteFunc } = require('./helpers/metatx');
-const { TLD, DEAD_ADDRESS } = require('./helpers/constants');
+const { TLD } = require('./helpers/constants');
 const { mintDomain } = require('./helpers/registry');
 
 const { utils, BigNumber } = ethers;
@@ -19,8 +19,7 @@ describe('UNSRegistry (metatx)', () => {
 
     unsRegistry = await UNSRegistry.deploy();
     await unsRegistry.initialize(coinbase.address);
-    await unsRegistry['mint(address,uint256,string)'](
-      DEAD_ADDRESS,
+    await unsRegistry['mintTLD(uint256,string)'](
       TLD.CRYPTO,
       'crypto',
     );
@@ -37,12 +36,7 @@ describe('UNSRegistry (metatx)', () => {
     const receiverAddress = '0x1234567890123456789012345678901234567890';
 
     it('should transfer using meta-setOwner', async () => {
-      const tokenId = await mintDomain(
-        unsRegistry,
-        owner,
-        TLD.CRYPTO,
-        'res_label_113a',
-      );
+      const tokenId = await mintDomain(unsRegistry, owner, ['res_label_113a', 'crypto']);
 
       const { req, signature } = await buildExecuteParams(
         'setOwner(address,uint256)',
@@ -56,12 +50,7 @@ describe('UNSRegistry (metatx)', () => {
     });
 
     it('should revert transfer using meta-setOwner when nonce invalidated', async () => {
-      const tokenId = await mintDomain(
-        unsRegistry,
-        owner,
-        TLD.CRYPTO,
-        'res_label_0896',
-      );
+      const tokenId = await mintDomain(unsRegistry, owner, ['res_label_0896', 'crypto']);
 
       const { req, signature } = await buildExecuteParams(
         'setOwner(address,uint256)',
@@ -104,12 +93,7 @@ describe('UNSRegistry (metatx)', () => {
     });
 
     it('should transfer using meta-transferFrom', async () => {
-      const tokenId = await mintDomain(
-        unsRegistry,
-        owner,
-        TLD.CRYPTO,
-        'meta_1591',
-      );
+      const tokenId = await mintDomain(unsRegistry, owner, ['meta_1591', 'crypto']);
 
       const { req, signature } = await buildExecuteParams(
         'transferFrom(address,address,uint256)',
@@ -123,12 +107,7 @@ describe('UNSRegistry (metatx)', () => {
     });
 
     it('should revert meta-transferFrom for non-onwer', async () => {
-      const tokenId = await mintDomain(
-        unsRegistry,
-        owner,
-        TLD.CRYPTO,
-        'meta_6458',
-      );
+      const tokenId = await mintDomain(unsRegistry, owner, ['meta_6458', 'crypto']);
 
       const { req, signature } = await buildExecuteParams(
         'transferFrom(address,address,uint256)',
@@ -143,12 +122,7 @@ describe('UNSRegistry (metatx)', () => {
     });
 
     it('should transfer using meta-safeTransferFrom', async () => {
-      const tokenId = await mintDomain(
-        unsRegistry,
-        owner,
-        TLD.CRYPTO,
-        'meta_10235',
-      );
+      const tokenId = await mintDomain(unsRegistry, owner, ['meta_10235', 'crypto']);
 
       const { req, signature } = await buildExecuteParams(
         'safeTransferFrom(address,address,uint256)',
@@ -162,12 +136,7 @@ describe('UNSRegistry (metatx)', () => {
     });
 
     it('should revert meta-safeTransferFrom for non-onwer', async () => {
-      const tokenId = await mintDomain(
-        unsRegistry,
-        owner,
-        TLD.CRYPTO,
-        'meta_e5iuw',
-      );
+      const tokenId = await mintDomain(unsRegistry, owner, ['meta_e5iuw', 'crypto']);
 
       const { req, signature } = await buildExecuteParams(
         'safeTransferFrom(address,address,uint256)',
@@ -184,12 +153,7 @@ describe('UNSRegistry (metatx)', () => {
     // TODO: add tests for safeTransferFrom(address,address,uint256,bytes)
 
     it('should burn using meta-burn', async () => {
-      const tokenId = await mintDomain(
-        unsRegistry,
-        owner,
-        TLD.CRYPTO,
-        'meta_ar093',
-      );
+      const tokenId = await mintDomain(unsRegistry, owner, ['meta_ar093', 'crypto']);
       const { req, signature } = await buildExecuteParams(
         'burn(uint256)',
         [tokenId],
@@ -204,12 +168,7 @@ describe('UNSRegistry (metatx)', () => {
     });
 
     it('should revert meta-burn for non-onwer', async () => {
-      const tokenId = await mintDomain(
-        unsRegistry,
-        owner,
-        TLD.CRYPTO,
-        'meta_53dg3',
-      );
+      const tokenId = await mintDomain(unsRegistry, owner, ['meta_53dg3', 'crypto']);
       const { req, signature } = await buildExecuteParams(
         'burn(uint256)',
         [tokenId],
@@ -263,10 +222,7 @@ describe('UNSRegistry (metatx)', () => {
       };
 
       const included = [
-        'mint(address,uint256,string)',
-        'safeMint',
-        'mintWithRecords',
-        'safeMintWithRecords',
+        'issueWithRecords',
       ];
 
       const getFuncs = () => {
@@ -286,9 +242,9 @@ describe('UNSRegistry (metatx)', () => {
       it('should execute all functions successfully', async () => {
         for (const func of getFuncs()) {
           const funcSigHash = utils.id(`${funcFragmentToSig(func)}_excl`);
-          paramValueMap.tokenId = await unsRegistry.childIdOf(
-            TLD.CRYPTO,
-            funcSigHash,
+          paramValueMap.tokenId = await unsRegistry.namehash(
+            [
+              funcSigHash, 'crypto'],
           );
 
           const req = await buidRequest(
@@ -321,12 +277,12 @@ describe('UNSRegistry (metatx)', () => {
       };
 
       const excluded = [
+        'mintTLD',
         'mint',
-        'safeMint',
-        'mintWithRecords',
-        'safeMintWithRecords',
+        'issue',
+        'issueWithRecords',
+        'unlockWithRecords',
         'transferFromFor',
-        'safeTransferFromFor',
         'burnFor',
         'resetFor',
         'setFor',
@@ -362,12 +318,7 @@ describe('UNSRegistry (metatx)', () => {
       it('should execute all functions successfully', async () => {
         for (const func of getFuncs()) {
           const funcSigHash = utils.id(`${funcFragmentToSig(func)}_ok`);
-          paramValueMap.tokenId = await mintDomain(
-            unsRegistry,
-            owner,
-            TLD.CRYPTO,
-            funcSigHash,
-          );
+          paramValueMap.tokenId = await mintDomain(unsRegistry, owner, [funcSigHash, 'crypto']);
 
           const req = await buidRequest(
             func,
@@ -389,12 +340,7 @@ describe('UNSRegistry (metatx)', () => {
         for (const func of getFuncs()) {
           const funcSig = funcFragmentToSig(func);
           const funcSigHash = utils.id(`${funcSig}_doubleUse`);
-          paramValueMap.tokenId = await mintDomain(
-            unsRegistry,
-            owner,
-            TLD.CRYPTO,
-            funcSigHash,
-          );
+          paramValueMap.tokenId = await mintDomain(unsRegistry, owner, [funcSigHash, 'crypto']);
 
           const req = await buidRequest(
             func,
@@ -420,12 +366,7 @@ describe('UNSRegistry (metatx)', () => {
         for (const func of getFuncs()) {
           const funcSig = funcFragmentToSig(func);
           const funcSigHash = utils.id(`${funcSig}_nonceInvalidated`);
-          paramValueMap.tokenId = await mintDomain(
-            unsRegistry,
-            owner,
-            TLD.CRYPTO,
-            funcSigHash,
-          );
+          paramValueMap.tokenId = await mintDomain(unsRegistry, owner, [funcSigHash, 'crypto']);
 
           const req = await buidRequest(
             func,
@@ -454,17 +395,9 @@ describe('UNSRegistry (metatx)', () => {
         for (const func of getFuncs()) {
           const funcSig = funcFragmentToSig(func);
           const funcSigHash = utils.id(`${funcSig}_wrongToken`);
-          paramValueMap.tokenId = await mintDomain(
-            unsRegistry,
-            owner,
-            TLD.CRYPTO,
-            funcSigHash,
-          );
+          paramValueMap.tokenId = await mintDomain(unsRegistry, owner, [funcSigHash, 'crypto']);
 
-          const tokenIdForwarder = await unsRegistry.childIdOf(
-            TLD.CRYPTO,
-            utils.id(`_${funcSig}`),
-          );
+          const tokenIdForwarder = await unsRegistry.namehash([utils.id(`_${funcSig}`), 'crypto']);
           const req = await buidRequest(
             func,
             owner.address,
@@ -489,12 +422,7 @@ describe('UNSRegistry (metatx)', () => {
           const funcSigHash = utils.id(
             `${funcFragmentToSig(func)}_emptyTokenId`,
           );
-          paramValueMap.tokenId = await mintDomain(
-            unsRegistry,
-            owner,
-            TLD.CRYPTO,
-            funcSigHash,
-          );
+          paramValueMap.tokenId = await mintDomain(unsRegistry, owner, [funcSigHash, 'crypto']);
 
           const req = await buidRequest(func, owner.address, 0, paramValueMap);
 
@@ -513,7 +441,7 @@ describe('UNSRegistry (metatx)', () => {
 
     describe('Non-Token functions', () => {
       const paramValueMap = {
-        label: 'label',
+        labels: ['label', 'crypto'],
         data: '0x',
         role: '0x1000000000000000000000000000000000000000000000000000000000000000',
         key: 'key_nt1',
@@ -558,7 +486,7 @@ describe('UNSRegistry (metatx)', () => {
       it('should execute all functions successfully', async () => {
         for (const func of getFuncs()) {
           const funcSig = funcFragmentToSig(func);
-          paramValueMap.label = utils.id(`${funcSig}_label`);
+          paramValueMap.labels = [utils.id(`${funcSig}_label`), 'crypto'];
 
           const req = await buidRequest(
             func,
@@ -579,13 +507,10 @@ describe('UNSRegistry (metatx)', () => {
       it('should revert execution of all functions when used signature', async () => {
         for (const func of getFuncs()) {
           const funcSig = funcFragmentToSig(func);
-          paramValueMap.label = utils.id(`${funcSig}_doubleUse`);
+          paramValueMap.labels = [utils.id(`${funcSig}_doubleUse`), 'crypto'];
           paramValueMap.key = utils.id(`${funcSig}_doubleUse`);
 
-          const tokenIdForwarder = await unsRegistry.childIdOf(
-            TLD.CRYPTO,
-            utils.id(`_${funcSig}`),
-          );
+          const tokenIdForwarder = await unsRegistry.namehash([utils.id(`_${funcSig}`), 'crypto']);
           const req = await buidRequest(
             func,
             coinbase.address,
@@ -609,7 +534,7 @@ describe('UNSRegistry (metatx)', () => {
       it('should revert execution of all functions when used signature and tokenId is empty', async () => {
         for (const func of getFuncs()) {
           const funcSig = funcFragmentToSig(func);
-          paramValueMap.label = utils.id(`${funcSig}_doubleUse_0`);
+          paramValueMap.labels = [utils.id(`${funcSig}_doubleUse_0`), 'crypto'];
 
           const tokenId = 0;
           const nonce = await unsRegistry.nonceOf(tokenId);